refactor(patientflowboard): move event subscriber out of zend_modules
