+++ src/Services/PatientFlowBoard/Subscriber/PatientFlowBoardEventsSubscriber.php
@@ -13,7 +13,7 @@
  * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
  */
 
-namespace OpenEMR\ZendModules\PatientFlowBoard\Listener;
+namespace OpenEMR\Services\PatientFlowBoard\Subscriber;
 
 use OpenEMR\Core\OEGlobalsBag;
 use OpenEMR\Events\Services\ServiceSaveEvent;
